refactor(reboot): unify the two reboot-progress dialogs onto the Vue RebootDialog
Two divergent reboot-wait + progress implementations existed: the raw-DOM
showRebootDialog() in serial_backend (used by CLI/OSD/MSPHelper) and a duplicate
wait-loop + Vue RebootDialog in useReboot() (used by 4 Vue tabs). The useReboot
copy called reinitializeConnection(true) to suppress the DOM one, ran its own
waitForReconnection, and — on the serial path — never called concludeReboot(),
leaving the phase stuck at REBOOTING (isReconnecting stuck true, which gates
selectActivePort). That was the "stuck REBOOTING" bug.

Consolidate on one implementation:
- showRebootDialog()/closeRebootDialog() now drive the shared Vue RebootDialog via
  the dialog store instead of a hand-built <dialog> element; createRebootProgressDialog
  and its inline styles are deleted.
- useReboot() reduces to a thin wrapper over reinitializeConnection() — no duplicate
  wait-loop, dialog, or suppressDialog=true. Every reboot now flows through the one
  backend path, which always settles the phase (fixing the stuck-REBOOTING case), and
  BLE/manual reboots from Vue tabs now show the same dialog they already show from CLI/OSD.
- Guard onClosed(): a reboot's own port-drop lands there, so it must NOT dismiss an
  active RebootDialog (the reboot flow owns it); non-reboot modals still get cleared.

Tests: RebootDialog-guard cases in serial_backend.test.js; existing reboot-parity and
BLE save-and-reboot suites still green. Reboot modal rendering wants a runtime/visual
check (same posture as the branch's other hardware-pending items).
